+++ Sources/KeyboardKit/Appearance/KeyboardStyle.swift
@@ -1,0 +1,14 @@
+//
+//  KeyboardStyle.swift
+//  KeyboardKit
+//
+//  Created by Daniel Saidi on 2023-08-03.
+//  Copyright © 2023 Daniel Saidi. All rights reserved.
+//
+
+import Foundation
+
+/**
+ This is a namespace for various keyboard styles.
+ */
+public struct KeyboardStyle {}