+++ Sources/KeyboardKit/System/SystemKeyboard.swift
@@ -341,14 +341,14 @@ public struct SystemKeyboard<ButtonView: View>: View {
     public typealias KeyboardWidth = CGFloat
     public typealias KeyboardItemWidth = CGFloat
 
-    private var actionCalloutStyle: KeyboardActionCalloutStyle {
+    private var actionCalloutStyle: KeyboardStyle.ActionCallout {
         var style = appearance.actionCalloutStyle
         let insets = layoutConfig.buttonInsets
         style.callout.buttonInset = CGSize(width: insets.leading, height: insets.top)
         return style
     }
 
-    private var inputCalloutStyle: KeyboardInputCalloutStyle {
+    private var inputCalloutStyle: KeyboardStyle.InputCallout {
         var style = appearance.inputCalloutStyle
         let insets = layoutConfig.buttonInsets
         style.callout.buttonInset = CGSize(width: insets.leading, height: insets.top)