+++ Sources/KeyboardKit/_Deprecated/KeyboardStyle+Deprecated.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+@available(*, deprecated, renamed: "KeyboardStyle.ActionCallout")
+public typealias KeyboardActionCalloutStyle = KeyboardStyle.ActionCallout
+
+@available(*, deprecated, renamed: "KeyboardStyle.Callout")
+public typealias KeyboardCalloutStyle = KeyboardStyle.Callout
+
+@available(*, deprecated, renamed: "KeyboardStyle.InputCallout")
+public typealias KeyboardInputCalloutStyle = KeyboardStyle.InputCallout